Se ajusta y modulariza css
