+++ Frontend/art-gallery/src/components/Filters/Filters.css
@@ -102,27 +102,6 @@ label {
   padding: 0.2em;
 }
 
-.text-range label {
-  font-size: 0.9em;
-}
-
-.subcontainer-range{
-  height: 100%;
-}
-.text-range {
-  height: 75%;
-}
-
-.text-range input {
-  height: 20%;
-  margin-top: 0.2em;
-  margin-bottom: 0.2em;
-  padding: 0.4em;
-  outline: none;
-  border: 2px solid var(--fifth-color);
-  border-radius: 8px;
-}
-
 .container-range label,
 .container-range input {
   color: var(--eigth-color);
@@ -130,35 +109,6 @@ label {
 
 /* Checkout List*/
 
-.subcontainerartist {
-  height: 100%;
-}
-
-.search-artist {
-  margin: 0.8em 0;
-  height: 12%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.search-artist input {
-  width: 90%;
-  height: 60%;
-  padding: 0.2em 0.4em;
-  border-radius: 12px;
-  border: 2px solid var(--fifth-color);
-  outline: none;
-}
-
-.artists {
-  height: 78%;
-  display: flex;
-  flex-direction: column;
-  border: 2px solid var(--fifth-color);
-  border-radius: 2px;
-  overflow-y: scroll;
-}
 
 .listArtist {
   padding: 0.07em 0.2em;
@@ -253,12 +203,6 @@ label {
   background-color: var(--primary-color);
 }
 
-
-/* Range */
-.range-slider{
-  width: 100%;
-  background-color: aqua;
-}
