+++ Frontend/art-gallery/src/components/utils/ListArtist/ListArtist.css
@@ -1,0 +1,30 @@
+.subcontainerartist {
+    height: 100%;
+  }
+
+  .search-artist {
+    margin: 0.8em 0;
+    height: 12%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .search-artist input {
+    width: 90%;
+    height: 60%;
+    padding: 0.2em 0.4em;
+    border-radius: 12px;
+    border: 2px solid var(--fifth-color);
+    outline: none;
+  }
+
+  
+.artists {
+    height: 78%;
+    display: flex;
+    flex-direction: column;
+    border: 2px solid var(--fifth-color);
+    border-radius: 2px;
+    overflow-y: scroll;
+  }
